[ci] use base sha in clang-tidy instead of base ref
The clang-tidy script diffs the PR head ref against the base ref so that
it works only on changed lines. If the base ref is a stale `master`,
then the script will fetch upstream `master` and potentially report
unrelated changes in the diff

Use the base sha instead of ref so that the revision that the script
diffs against is stable.

ghstack-source-id: c084f3f77bc125e787d61b12011c9befaff5053a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28388