Stop using RTLD_GLOBAL for _C
This should help us resolve a multitude of weird segfaults and crashes
when PyTorch is imported along with other packages. Those would often
happen because libtorch symbols were exposed globally and could be used
as a source of relocations in shared libraries loaded after libtorch.

One complication is that MKL doesn't like being in a local scope, unless
you use the libmkl_rt.so variant, so I adapted our CMake configuration
to either accept that or only allow static linking.

All relocations are also resolved eagerly now, which should let us avoid
some bugs. In the process of writing this patch I found e.g. a missing
TORCH_API declaration in the NCCL bindings.

We also no longer use ctypes to interface with CUDA libraries, but use
pybind to expose the relevant functionality. This has the upside of
no longer forcing us to hardcode the enum values in Python, and also
should let us remove the ungodly hacks we use to load the libraries on
Windows.

Finally, turns out that most of our cuDNN interface in Python is dead
now so I have removed it.